Add translations for some core open source process files
This addresses #37. Note that these translations were requested from the
State Dept a few weeks ago, so they don't have some of the latest
updates to content from recent PRs. We will need to come up with a good
proces for keeping both sets of docs up to date as we make changes.