Add translations for repo documents (#281)
* WIP README improvements

* Add translations for some core open source process files

This addresses #37. Note that these translations were requested from the
State Dept a few weeks ago, so they don't have some of the latest
updates to content from recent PRs. We will need to come up with a good
proces for keeping both sets of docs up to date as we make changes.

Co-authored-by: Shelby Switzer <shelby.switzer@cms.hhs.gov>