Update install steps (#359)
* Update INSTALLATION.md

Update INSTALLATION with more narrative text on how to install and start running parts of the app. We will need to update this soon with more info about the repo organization (once we have a `data` directory), but for now this reflects the current state of the repo.

* Update docs

Move client-specific things to client README, and update client README with steps to get the clientside app running locally.