Fix putting elements under <p>
Added Jest test to catch this mistake in the future. It does it by watching console.error().
To add in fixing the bug the test's emulated `console.error()` are echoed to the jest console.error() so they are not lost.

I tested this by putting a `<h1>` back into the `<p>` and it caught it.