Add action for ping test and alert to Slack (#695)
* Add action for ping test and alert to Slack

Based on @sverchdotgov 's https://github.com/sverchdotgov/github-actions-experiments/blob/master/.github/workflows/scheduled-email.yml

This is set to run every hour. We can modify as needed. It is also using our cloudfront URL and we'll need to change this to our new URL once we have it.

* Update main.yml

Change to run every 5 minutes and update env var to be more descriptive based on PR convo