Remove "infrastucture" directory (#996)
We originally used the approach in the infrastructure directory when we
thought we would be using Amazon Lambda for different parts of our
deployment pipeline. We have since then moved to using Github Actions
and no longer need this code, and keeping it in `main` has caused
confusion for onboarding new folks. This commit removes the directory
(although this will still be around in version control so we can always
view it or bring it back in the future if we want to).

Co-authored-by: Shelby Switzer <shelby.switzer@cms.hhs.gov>