Adding github actions to test score and score post
In service of https://github.com/usds/justice40-tool/issues/599

This does not test the etl jobs specific to the datasources. Instead, it
takes the "datasets" directory created after all those run and uploads
it to s3 in one github action, and downloads it in another. We believe
this will take about 8 minutes to run end to end based on
https://github.com/usds/justice40-tool/pull/1039#pullrequestreview-830528432.