Added pandas numeric conversion example
Trying to find a sane way to convert the type of a column. It seems like
creating a new Series object with a "dtype" argument does it and
preserves NaNs.

Also added jupyterlab as a dev dependency. The best advice I ever got on
pandas is to struggle with it in an ipython notebook until something
makes sense. Sure enough, this took like 30 attempts, but the quick
turnaround time helped and I think this approach works.